Improve ceph tools: support arbitrary nodes number, support raw/loop device. (#427)

+++ deployment/ceph-tools/deploy/stop-all.sh
@@ -1,20 +1,31 @@
 set -eu
 source ./_env.sh
 
-for node in ${nodes[@]}; do
+for i in ${!nodes[@]}; do
+	if [ $i -gt 2 ]; then
+		continue
+	fi
+	node=${nodes[$i]}
 	ssh "$node" "sudo systemctl stop ceph-mds@$node"
 done
 
-for node in ${nodes[@]}; do
+for i in ${!nodes[@]}; do
+	if [ $i -gt 2 ]; then
+		continue
+	fi
+	node=${nodes[$i]}
 	ssh "$node" "sudo systemctl stop ceph-mgr@$node"
 done
 
-for node in ${nodes[@]}; do
+for i in ${!nodes[@]}; do
+	if [ $i -gt 2 ]; then
+		continue
+	fi
+	node=${nodes[$i]}
 	ssh "$node" "sudo systemctl stop ceph-mon@$node"
 done
 
-for i in ${#nodes[@]}; do
-	i=$(( $i - 1 ))
+for i in ${!nodes[@]}; do
 	ssh "${nodes[$i]}" "sudo systemctl stop ceph-osd@$i"
 done
